% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate_ddl.R
\name{build_column_ddl}
\alias{build_column_ddl}
\title{Build Column DDL SQL Fragment}
\usage{
build_column_ddl(field, strategy, target)
}
\arguments{
\item{field}{Character scalar. Column name.}

\item{strategy}{Character scalar. Transformation strategy.}

\item{target}{Character scalar. Target PostgreSQL type where applicable.}
}
\value{
Character scalar column DDL SQL fragment.
}
\description{
Generates SQL fragment defining a column name and PostgreSQL type.
}
