% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promote_ddl.R
\name{build_join_clause_vector}
\alias{build_join_clause_vector}
\title{Build Vector of JOIN Clauses from Schema}
\usage{
build_join_clause_vector(schema, build_join_sql, table_name)
}
\arguments{
\item{schema}{Data frame or list compatible with purrr::pmap.}

\item{build_join_sql}{Function used to build join SQL.}

\item{table_name}{Character scalar. Destination table name.}
}
\value{
Character vector of SQL JOIN clauses.
}
\description{
Applies join builder across schema definition and removes NA entries
where joins are not required.
}
