% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_ddl.R
\name{build_load_script_header}
\alias{build_load_script_header}
\title{Build shell script header for loading raw data into landing schema}
\usage{
build_load_script_header(file_name, file_folder)
}
\arguments{
\item{file_name}{Name of the raw data file.}

\item{file_folder}{Folder under \verb{$DATASTORE} containing the file.}
}
\value{
Character string containing bash script header.
}
\description{
Generates the initial part of a bash script that:
\itemize{
\item Defines the INPUT file location
\item Checks that the file exists before continuing
}
}
