% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promote_ddl.R
\name{build_promote_function_sql}
\alias{build_promote_function_sql}
\title{Build PL/pgSQL Import Function SQL}
\usage{
build_promote_function_sql(selects, joins, schema, table)
}
\arguments{
\item{selects}{Character vector of SELECT expressions.}

\item{joins}{Character vector of JOIN clauses.}

\item{schema}{Character scalar. Destination schema name.}

\item{table}{Character scalar. Destination table name.}
}
\value{
Character scalar SQL function definition.
}
\description{
Generates SQL text for a PL/pgSQL function that imports data from staging
into a destination schema table.
}
