% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promote_ddl.R
\name{build_select_clause_vector}
\alias{build_select_clause_vector}
\title{Build Vector of SELECT Expressions from Schema}
\usage{
build_select_clause_vector(schema, build_column_select_sql)
}
\arguments{
\item{schema}{Data frame or list compatible with purrr::pmap.}

\item{build_column_select_sql}{Function used to build column SQL expressions.}
}
\value{
Character vector of SQL SELECT expressions.
}
\description{
Applies column expression builder across schema definition to produce
a vector of SQL SELECT expressions.
}
