% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate_ddl.R
\name{build_table_ddl}
\alias{build_table_ddl}
\title{Build PostgreSQL CREATE TABLE DDL SQL}
\usage{
build_table_ddl(schema_tbl, table_name, pg_schema = "dft")
}
\arguments{
\item{schema_tbl}{Data frame or list compatible with purrr::pmap,
containing column mapping definitions.}

\item{table_name}{Character scalar. Table name to create.}

\item{pg_schema}{Character scalar. PostgreSQL schema name.}
}
\value{
Character scalar CREATE TABLE SQL statement.
}
\description{
Generates CREATE TABLE SQL using a schema mapping table.
}
