% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum_builders.R
\name{generate_all_enum_drop_sql}
\alias{generate_all_enum_drop_sql}
\title{Generate drop enum code}
\usage{
generate_all_enum_drop_sql(df)
}
\arguments{
\item{df}{The name of th4e dataframe containing
table, field_name. The latter is the name of the enums to be
dropped.}
}
\value{
SQL
}
\description{
Given the same data frame (containing table, field_name and labels)
will create sql to drop these enums
}
