% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum_builders.R
\name{generate_all_enum_sql}
\alias{generate_all_enum_sql}
\title{Generate a set of enums}
\usage{
generate_all_enum_sql(df)
}
\arguments{
\item{df}{A data frame containing table, field_name and label}
}
\value{
A long string encoding all enums
}
\description{
Given a dataframe containing \code{table} and \code{field_name}
assuming one row for each value of the enum
This will create the sql to specify these enums.
Will only do this for coded variables with 3 or more and less than 15 values
}
