% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum_builders.R
\name{generate_enum_sql}
\alias{generate_enum_sql}
\title{Generate enum}
\usage{
generate_enum_sql(field, values, schema = "dft", prefix = "stats19")
}
\arguments{
\item{field}{The name of the variable to be enummed}

\item{values}{The values of the enum}

\item{schema}{The schema to store the enum}

\item{prefix}{the data prefix (group name)}
}
\value{
A multiline string which should be valid sql
}
\description{
Generate code to create an enum in Postgres
}
