% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promote_ddl.R
\name{generate_promotion_sql}
\alias{generate_promotion_sql}
\title{Generate Import SQL for Stats19 Table}
\usage{
generate_promotion_sql(
  intended_schema,
  build_column_select_sql,
  build_join_sql,
  table,
  psql_schema = "dft"
)
}
\arguments{
\item{intended_schema}{Schema definition used for column mapping.}

\item{build_column_select_sql}{Function used to build column SELECT SQL.}

\item{build_join_sql}{Function used to build JOIN SQL.}

\item{table}{Character scalar. Base table name (without prefix).}

\item{psql_schema}{Character scalar. Destination PostgreSQL schema.
Default is \code{"dft"}.}
}
\value{
Character scalar SQL definition for import function.
}
\description{
High-level orchestration function that builds SELECT clauses, JOIN clauses,
and final PL/pgSQL import function SQL for a given schema mapping.
}
