% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{guess_encoding}
\alias{guess_encoding}
\title{Guess File Text Encoding}
\usage{
guess_encoding(file)
}
\arguments{
\item{file}{Character scalar. Path to the file.}
}
\value{
A tibble describing possible encodings and confidence scores.
}
\description{
Wrapper around \code{readr::guess_encoding()} to detect likely text encodings
for an input file.
}
\details{
This function simply forwards to \code{readr::guess_encoding()} and exists
primarily for convenience and consistency within the package.
}
\examples{
\dontrun{
guess_encoding("data/collisions.csv")
}

}
\seealso{
readr::guess_encoding
}
