% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{write_schema_yaml}
\alias{write_schema_yaml}
\title{Write schema yaml}
\usage{
write_schema_yaml(schema_tbl, file, metadata = list())
}
\arguments{
\item{schema_tbl}{A tibble listing the looku pdata}

\item{file}{the target file name}

\item{metadata}{Important metadata we also want to store in the yaml}
}
\description{
Write schema yaml
Assuming we have a spec in a schema_tbl
}
\examples{
\dontrun{
write_schema_yaml(
 schema_casualty,
 "schemas/stats19_casualty.yml",
 metadata = list(
   dataset = "stats19_casualty",
   download_date = as.character(Sys.Date()),
   source = "DfT Stats19",
   notes = "Generated from 2024 metadata"
 )
)
}
}
