% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_ddl.R
\name{build_load_script_footer}
\alias{build_load_script_footer}
\title{Build shell script footer for loading and promoting data}
\usage{
build_load_script_footer(pg_table, year = 2024)
}
\description{
Generates the part of the bash script that:
\itemize{
\item Copies CSV data into landing table
\item Calls database promote function
}
}
